Move global strategies to subdirectory
