+++ lua/rainbow-delimiters/strategy/global/current.lua
