+++ lua/rainbow-delimiters/strategy/global/legacy.lua
